+++ tools/PikaLSteviappModule/gui/trajectoryvieweditor.h
@@ -98,8 +98,8 @@ protected:
     QSpinBox* _startLineSpinBox;
     QSpinBox* _endLineSpinBox;
 
-    std::vector<float> _baseTrajectoryTimes;
-    std::vector<float> _compareTrajectoryTimes;
+    std::vector<double> _baseTrajectoryTimes;
+    std::vector<double> _compareTrajectoryTimes;
 };
 
 class TrajectoryViewEditorFactory : public StereoVisionApp::EditorFactory
